fix: 再提出フローのレビュー指摘を反映

+++ api/app/controllers/power_orders_controller.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 class PowerOrdersController < ApplicationController
-  before_action :authenticate_api_user!, only: %i[get_by_group_id submit]
+  before_action :authenticate_api_user!
   before_action :set_power_order, only: [:show]
   before_action :set_power_orders_by_group_id, only: [:get_by_group_id]
 
   # GET /power_orders
   # GET /power_orders.json
   def index
-    @power_orders = PowerOrder.all
+    @power_orders = PowerOrder.where(group_id: current_api_user.groups.select(:id))
     render json: fmt(ok, @power_orders)
   end
 
@@ -26,12 +26,14 @@ class PowerOrdersController < ApplicationController
   def submit
     group = current_user_group
     return render_submit_not_found unless group
+    return render_submit_unprocessable_entity('use_power is required') unless params.key?(:use_power)
+    return render_submit_unprocessable_entity('power_orders is required') if use_power? && power_order_params_list.empty?
 
     ActiveRecord::Base.transaction do
       if use_power?
         delete_unregistered_power_order(group)
-        delete_missing_power_orders(group)
-        upsert_power_orders(group)
+        saved_ids = upsert_power_orders(group)
+        delete_missing_power_orders(group, saved_ids)
       else
         group.power_orders.destroy_all
         ensure_unregistered_power_order(group)
@@ -55,8 +57,12 @@ class PowerOrdersController < ApplicationController
 
   # Use callbacks to share common setup or constraints between actions.
   def set_power_order
-    if PowerOrder.exists?(params[:id])
-      @power_order = PowerOrder.find(params[:id])
+    @power_order = PowerOrder
+                   .where(group_id: current_api_user.groups.select(:id))
+                   .find_by(id: params[:id])
+
+    if @power_order
+      @power_order
     else
       render json: fmt(not_found, [], "Not found power_order = #{params[:id]}")
     end
@@ -76,9 +82,7 @@ class PowerOrdersController < ApplicationController
 
   # Only allow a list of trusted parameters through.
   def current_user_group
-    return nil if params[:group_id].blank?
-
-    current_api_user.groups.find_by(id: params[:group_id])
+    current_api_user_group(params[:group_id])
   end
 
   def use_power?
@@ -86,7 +90,7 @@ class PowerOrdersController < ApplicationController
   end
 
   def upsert_power_orders(group)
-    power_order_params_list.each do |power_order_params|
+    power_order_params_list.map do |power_order_params|
       power_order = if power_order_params[:id].present?
                       group.power_orders.find_by(id: power_order_params[:id])
                     else
@@ -97,13 +101,12 @@ class PowerOrdersController < ApplicationController
 
       power_order.assign_attributes(power_order_params.except(:id))
       power_order.save!
+      power_order.id
     end
   end
 
-  def delete_missing_power_orders(group)
-    submitted_ids = power_order_params_list.filter_map { |power_order| power_order[:id].presence }
-
-    group.power_orders.where.not(id: submitted_ids).destroy_all
+  def delete_missing_power_orders(group, saved_ids)
+    group.power_orders.where.not(id: saved_ids).destroy_all
   end
 
   def power_order_params_list
@@ -125,5 +128,9 @@ class PowerOrdersController < ApplicationController
 
   def render_submit_not_found
     render json: fmt(not_found, [], 'power_order not found'), status: :not_found
+  end
+
+  def render_submit_unprocessable_entity(message)
+    render json: fmt(unprocessable_entity, [], message), status: :unprocessable_entity
   end
 end
